refactor-search 패키지 오류 해결 및 네이밍 수정
