+++ src/main/java/com/project/Instagram/domain/search/repository/querydsl/RecentSearchRepositoryQuerydsl.java
@@ -1,0 +1,4 @@
+package com.project.Instagram.domain.search.repository.querydsl;
+
+public interface RecentSearchRepositoryQuerydsl {
+}
