+++ src/main/resources/static/websocket.html
@@ -7,6 +7,7 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/stomp.js/2.3.3/stomp.min.js"></script>
     <!-- SockJS 라이브러리 포함 -->
     <script src="https://cdnjs.cloudflare.com/ajax/libs/sockjs-client/1.5.1/sockjs.min.js"></script>
+    <!-- JavaScript 파일 포함 -->
     <script src="websocket.js"></script>
 </head>
 <body>
@@ -16,8 +17,6 @@
 <button onclick="deleteMessage()">Delete Message</button>
 <input type="text" id="messageInput" placeholder="Enter your message">
 <div id="messages"></div>
-
-<!-- JavaScript 파일 포함 -->
 
 </body>
 </html>
